api/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
alembic
pydantic
python-dotenv
python-jose
passlib[bcrypt]
